Fix: Extracted the sorting logic to a custom hook.
